Fix RejectedExecutionException when mirroring workers are busy
Motivation:

DefaultMirroringService uses a SynchronousQueue as the task queue of
mirroring worker thread pool so that the worker queue does not grow
infinitely when the workers cannot catch up the schedule.

However, ThreadPoolExecutor will fail to submit a mirroring task with a
RejectedExecutionException when all workers are busy, because a
SynchronousQueue does not have capacity.

When a RejectedExecutionException is triggered, DefaultMirroringService
will stop schedule mirroring completely.

Modifications:

- Specify RejectedExecutionHandler so that RejectedExecutionExeception
  is not triggered and the scheduler simply waits until the task is accepted
- Add a test case

Result:

MirroringService does not terminate when it cannot catch up.