Compress API responses if requested (#211)
Motivation:

An API response can be as big as the size of the entries stored in
Central Dogma. A user may want to reduce the traffic by sending his or
her request with a 'accept-encoding: deflate' (or gzip) header.

Modifications:

- Apply HttpEncodingService to Thrift and HTTP API endpoints

Result:

A user can choose to get a compressed response.