Do not list partially-created projects unless necessary
Motivation:

When we access a project, we usually assume that `dogma` and `meta`
repositories exist in that project. However, the creation of a project
is a multi-step process:

- Create a project
- Create the `dogma` repository
- Create the `meta` repository

Therefore, if another party who attempts to access the project (or its
`meta` repository) at the same time, he or she may see a partial view of
the project.

Modifications:

- Move the creation of `dogma` and `meta` repository to `DefaultProject`
  so that a project creation is fully transactional and thus a failed
  project creation will roll back the partial changes.
  - `ProjectInitializingCommandExecutor` is gone.
- Modify `StorageManager.create()` so that the storage subsystems can
  record the creation time and project author correctly.
- Miscellaneous:
  - Replace `INTERNAL_REPO` and `METADATA_REPO` with `Project.REPO_DOGMA`
  - Prevent API callers from creating `meta` or `dogma` repository,
    because they must be created by the system when a project is created.

Result:

- Nobody can see a partially created project.
- Fixes #245