Add 'ID' field to credentials
Motivations:

Sometimes, a user has to specify different credentials for different
repositories that reside in the same host. In this case, we cannot
select the credentials based on hostname patterns, because their
hostnames are same.

Modifications:

- Add `id` field to `credentials.json`
- Add `credentialId` and `defaultCredentialId` field to `mirrors.json`
- Update the documentation

Result:

Can specify different credentials for different repositories that reside
in the same host