Provide modularized authentication system
Motivation:
For now, Apache Shiro provides an authentication service for Central Dogma server. It's enough for LDAP, JDBC or Active Directory based authentication, but not useful for some authentication methods like OAuth or SAML. So it would be better to split the current authentication layer into abstraction authentication layer and its implementation layer so that we can replace the authenticaion implementation with the new one easily.

Modifications:

- Remove Apache Shiro from server dependencies, and make a new module.
- Add c.l.c.server.authentication package and provide AuthenticationProvider and AuthenticationProviderFactory interfaces for authentication implementations.
- Add session manager classes which have features provided by FileBasedSessionDAO before.
- Add saml module in order to provide SAML based authentication.
  - SamlAuthenticationProperties is a start point for configuring SAML authentication.
- Add webapp module in order to provide default web pages for login and logout.
  - An authentication system which provides a login and logout page by itself, such as Apache Shiro, can use the default web pages in the module.
  - SAML authentication system can also use the default logout page in the module.

Result:
A user can choose one of Apache Shiro or SAML as his or her authentication system.