Load client profiles from all resources
Motivation:

Currently, Central Dogma client uses only the first found
`centraldogma-profiles.json` file to load client profiles. It would be
useful to load all `centraldogma-profiles.json` files in the class path
so that it is possible for a user to add custom profiles on top of the
official profiles provided by administrators.

Modifications:

- Add `priority' property to a client profile
- Load all profile JSONs so that all profiles are considered.
- Make sure the profile that has higher priority or appeared first in
  the classpath wins when there are more than one profile with the same
  name.
- Update documentation

Result:

- Fixes #292