Interrupt Central Dogma server startup process if stop is requested.
Motivation:

In a replicated setup, the startup process of Central Dogma server can
take an extended period of time if the cluster does not contain the
initial quorum. In this situation, any attempt to stop the server will
block until the cluster contains the initial quorum and thus a user had
to terminate the process manually to stop the server.

Modifications:

- Stop waiting for the quorum if there is a pending stop request.
  - Made the following changes to `CentralDogmaRule` to test this scenario:
    - Made `CentralDogmaRule.before()` and `after()` non-final.
    - Added `CentralDogmaRule.startAsync()` and `stopAsync()`.
    - Made sure the temporary folder is deleted after the server stops.
- Miscellaneous:
  - Fixed intermittent `FileNotFoundException` in `ZooKeeperCommandExecutorTest`.
  - Renamed `ReplicationConfigTest` to `NoneReconfigurationConfigTest`.

Result:

- A user can stop a server even if it is stuck in the startup process.
- Less flakiness