Adjust timeout for HTTP `watch` requests according to `wait` parameter
Motivation:
Currently Central Dogma server does not adjust timeout for `watch` requests.
So every `watch` request would be timed out after the default timeout in the server even if the request contains `wait` value which is greater than the default timeout. It causes that the server responds `503 Service Unavailable` to the client.

Modifications:
- Adjust timeout for `watch` requests in `WatchRequestConverter`.
- Change the max timeout from `Long.MAX_VALUE` to `TimeUnit.DAYS.toMillis(1)` (1 day).
- Fix test cases.

Result:
- A user can adjust the timeout of a `watch` request.