Show configured title on the web-based administrative console (#349)
Motivation:
Currently, there's no easy way to know if the CD instance he or she is accessing belongs to BETA or RELEASE zone. We currently show the host name of the instance, but it's not clear enough. We could add a simple text value option whose value is appended to the web page title. e.g. `Central Dogma (BETA, {hostname})`

Modifications:
- Changed to send both of hostname and title from `/hostname` API.
  - The title will be displayed on the navigation bar of the web console.
- Updated the document about the configuration.

Result:
- Make web console users less confused.