Make webAppTitle fully customizable
Motivation:
A user could configure the title of web-based administrative console by #348. But it is only available to add some text to the end of pre-defined title which is `Central Dogma at ..hostname..`. We think it would be more useful if a user can customize the entire title of the web console.

Modifications:
- Made the entire title of the web console customizable.
  - A user can use ``{{hostname}}`` as a variable of the actual hostname that the server is running on.
- Updated site document.
- Misc
  - Try to move pidfile again silently when it caught `AtomicMoveNotSupportedException`.

Result:
- Closes #43
- Usability.