Add `revision` property to `EntryDto`
Motivation:

We currently do not return the revision of an entry. The reasoning
behind this behavior is that the response would contain duplicate
'revision' fields if a user requested more than one entry. Because of
that, there's a problem that a user could not notice quickly which
version of a file is loaded into his/her service.

Modifications:

- Added `revision` property to `EntryDto`
- Miscellaneous:
  - Used `HttpHeaderNames.PREFER` which reappeared Armeria 0.83.0
  - Used `HttpData.toStringUtf8()` instead of `.content().toStringUtf8()`

Result:

- Closes #310