Retry client requests on replication lag
Motivation:

A client sometimes gets different entries from different replicas in the
same Central Dogma cluster, mainly due to replication lag, i.e. Even if
you learned that the latest revision of a repository in replica A is 42,
you may fail to get a file at revision 42 from replica B.

Currently, in such a case, our Java client simply fails with
`RevisionNotFoundException`, which is confusing to our users.

We could instead keep the latest known revision numbers of the
repositories and retry a few more times when it is certain that a
`RevisionNotFoundException` was raised due to the replication lag.

Modifications:

- Make all shortcut methods in `CentralDogma` final in
  `AbstractCentralDogma` so that a dev can focus on less number of
  methods.
- Add `ReplicationLagTolerantCentralDogma`
- Add the properties related with replication lag tolerance to
  `AbstractCentralDogmaBuilder`:
  - `maxRetriesOnReplicationLag`
  - `retryIntervalOnReplicationLag`

Result:

- Replication lag is detected and the request is retried automatically.
- Fixes #435