Do not run `GitAuthTest` by default
Motivation:

`GitAuthTest` requires you to have an access to a private repository
which is only accessible by LINE employees by default. It currently runs
when a user has an SSH key pair in `~/.ssh/id_rsa{,.pub}` regardless
whether he or she has an access to the repository, leading to a test
failure.

Modifications:

- The test now requires a user to set `git.privateKey` System property
  explicitly to run the test.

Result:

- Build passes anywhere.