Use a blocking task executor for the watch callacks of Central Dogma client
Motivation:

The callacks for watcher are synchronous APIs such as `Function<T, U>`.
Users can not compose an asynchronous API for the callback function.
They have to wait for the result and return it.
```java
dogma.repositoryWatcher("xyz", "foo", "/*.txt",
                        revision -> {
                            dogma.getFiles("xyz", "foo",
                                           revision,
                                           "/*.txt").join();
                            return "Whatever";
                        });
```

CentralDomga internally invokes the callback by an event loop which must not be blocked.
However, most users do not aware that the callback is called by an event loop.
It would be safe to call the callbacks of watcher with a blocking task executor.

Modifications:

- Replace `EventLoop` executor with a blocking task executor.

Result:

- You can now safely execute a block IO for the callbacks of watcher.
- Fixes #587