Allow configuring the initialization timeout of `CentralDogma` client for Spring integration. (#692)
Motivation:

3.2 seconds is used for the default timeout if an endpoint is not
determined while executing a request.
While a server is starting up, a resolution of endpoints may take
longer than 3.2 seconds because the server could be busy initializing
the application context.
To solve the problem, `CentralDogma.whenEndpointReady()` was introduced
to wait for the initial endpoints before starting a request.

However, users found it easy to forget to call the API and use directly a
`CentralDogma` client without waiting for it. It is difficult to call
`whenEndpointReady().get()` because it is to block the current thread.
But a `CentralDogma` bean is a managed instance created in the main
thread when a Spring application context is initialized. So we can
safely call `CentralDogma.whenEndpointReady().get()` before returning a
client.

Modifications:

- Add `initializationTimeoutMillis` to `CentralDogmaSettings` so as to
  to let users configure the timeout
  - If unspecified, 15 seconds is used by default.
  - 0 disables the waiting for initialization.

Result:

- You can now configure the initialization timeout for a `CentralDogma`
  client when using Spring integration modules.
  ```yml
  centraldogma:
    hosts:
       - ...
    initialization-timeout-millis: 15000
  ```
- Fixes #691