Add a trailing slash to the auth URLs
Motivation:

Next.js generates index.html files under the target paths when
`trailingSlash` option is enabled. Additionally, `/` is automatically
added to a routing path.
https://nextjs.org/docs/api-reference/next.config.js/trailing-slash

Note:

This PR is a temporary workaround to respond to trailing slashs properly.
The URLs will be reverted and `trailingSlash` option is disabled when
https://github.com/line/armeria/issues/4542 is resolved.

Modifications:

- Appended `/` to URLs related to authentication.
- Remove the legacy auth file service to avoid path conllisions.

Result:

- Users can login with Shiro authentication.