Improve the error message of a `TimeoutException` from `watcher.await… (#750)
…InitialValue()`

Motivation:
A `TimeoutException` is raised when using `watcher.awaitInitialValue(...)` if the target entry doesn't exist. We can give a hint that uses `errorOnEntryNotFound` in the exception message so that users who don't know the flag can give it a try if it's the problem.

Modifications:
- Provide a detailed error message when `Watcher.awaitInitialValue(...)` raises `TimeoutException`.

Result:

You can see the detailed error message when `Watcher.awaitInitialValue(...)` raises a `TimeoutException` because the target doesn't exist.