Extract mirroring implementation from the server module (#836)
Motivation:
Previously, the server module was using jgit5 and jgit6 libraries together. It's because jgit6 requires JDK 11 as a minimum, so we needed jgit5 for JDK 8 client.
However, that might cause a dependency conflict problem such as #835. If we extract jgit6 specific API and implementations to a new module, we can avoid this problem.

Modifications:
- Add `MirrorProvider` interface that creates a `Mirror`.
  - `MirrorProvider`s are discovered via SPI.
- Created a new module called `server-mirror-git` and added `GitMirrorProvider`.
- jgit6 is used if `javaTestVersion` is greater or equal to 11.
  - jgit5 is sued if `javaTestVersion` is 8.
- Clean up dependencies in build.gradle.
- (Breaking)
  - The `type` property, which was useless, in the mirroring configuration is now removed.
  - Removed `server-git6` module.

Result:
- The server module is now correctly built and tested with the proper JGit libraries depending on the JDK test version.