Check remote revision before fetching when mirroring (#881)
Motivation:
If there are no commits from the last mirroring, the mirroring service does not have to fetch from the remote.
Because commits are infrequently made, we can just use `git-ls-remote`, which is way cheaper than `git-fetch-pack`, to check the remote head revision.

When mirroring, an SSH session might be established a couple of times. Establishing an SSH session is a highly resource-intensive task. It would be nice if we could reduce it.

Modifications:
- Check the remote revision before fetching when mirroring if there's a new commit.
- Split `GitMirror` into:
  - `AbstractGitMirror` that is the common ancestor of `GitMirror` and `SshGitMirror`.
  - `SshGitMirror` for SSH mirroring and `GitMirror` for the rest of them.
- Share SSH client and session for a mirror job.

Result:
- Partially fix #879
- The average mirroring duration is now decreased approximately by 50%.