Fix to show project creators correctly (#909)
Motivation:
The current creator of a project is the creator of the internal `dogma` repository, consistently set as `System`. We should show the correct creator which is stored in the `metadata.json` file.

A noteworthy observation is that only the email of the creator is stored in `metadata.json` instead of the complete `Author` information. A temporary solution involves restoring the `Author` using the email, with plans to revisit this approach if issues arise.

Modifications:
- Fix to show project creators correctly using information stored in the `metadata.json` file.

Result:
- Close #908 
- The project creator is now correctly retrieved based on the information stored in the `metadata.json` file.