Enable admin access to the internal `dogma` project. (#907)
Motivation:
@trustin suggested utilizing the internal `dogma` project for the
control plane. However, it's not easy to use the project because it
isn't accessible through UI even for an admin.
The `ProjectManager`, which is used in the `ProjectServiceV1`, is
wrapped with the `SafeProjectManager` that blocks to access the `dogma`
project. (A Plugin uses the `ProjectManager` directly so no problem from
the plugin side.)

Modifications:
- Grant admins to access to the internal `dogma` project via REST API
and web UI.
- Rename `SafeProjectManager` to `ProjectApiManager`.
  - It's also used for Thrift service.

Result:
- Administrators can now access the internal `dogma` project through
both REST API and the web UI.