Merge pull request from GHSA-34q3-p352-c7q8
Motivation:
SAML `RelayState` is utilized to direct authenticated users to the correct resource. As it is inserted into the HTML document, proper HTML-escaping is crucial to prevent potential security risks, such as XSS attacks.
[Reference](https://developer.okta.com/docs/concepts/saml/)
For example, when the `RelayState` is `'.substr(0,1)...`, the redirection script will be:
`<html ... <body onload="window.location.href='/#'.substr(0,1)....."></body></html>` which an attacker can insert XSS code to `.....`.

Modification:
- Implement HTML-escaping for SAML `RelayState` when creating the redirection script.

Result:
- Improved security by mitigating the risk of XSS attacks through HTML-escaping of SAML `RelayState`.