Sleep 1 second after changing server healthy to false when shutting down (#947)
We changed to sleep as much as graceful shutdown quiet period after changing server healthy when shutting down in https://github.com/line/centraldogma/pull/938
However, if the value is too large, it may result in an unnecessarily long wait. Let's just change the value to 1 second and add a parameter later if there's a demand.