Restrict prefix value in `CentralDogmaConfig`
Motivation:
The current implementation of `CentralDogmaConfig` strips the prefix separated by a colon from the configuration value.
This causes issues when the configuration value, without a prefix, contains a colon.
For instance, the following configuration value:
```
-----BEGIN EC PRIVATE KEY-----
Proc-Type: 4,ENCRYPTED
DEK-Info: ...
...
```
In this case, `CentralDogmaConfig` throws an exception because there is no `ConfigValueConverter` for converting the value after the colon.
This can be avoided by restricting the prefix value to only valid characters.

Modifications:
- Restrict the prefix value to the regular expression pattern `^[a-z0-9_]+$`.
- Use the configuration value as is if `CentralDogmaConfig` fails to find a proper `ConfigValueConverter` for the prefix.

Result:
- `CentralDogmaConfig` will no longer fail due to configuration values with colons.