Add a metric for tracking the number of pushes (#953)
Motivation:

It isn't easy to check how many commits have been pushed to a specific repository without looking at log files.

This PR is the orignal commit of #943 that tracks the number of pushes instead of adding `project` and `repo` tags to all `api` metrics. #943 has been reverted because it increased in CPU and memory usage.

Modifications:

- Add a `Counter` that is increased when a commit is pushed to a repository.
  - `project` and `repository` are added as tags.

Result:

You can monitor the number of pushes to a repository using the `commits.push` metric.