Wait for internal projects to be created before starting the mirror migration
Motivation:

Plugins are started asynchronously when a command executor starts.
A mirror migration job needs to access `dogma/dogma` project to check if
a migration job was performed before.

I worked a similar logic before. After not working on it for a long
time, I deleted the code when merging main branch. I was confused and I
thought it was redudunt.

Modifications:

- Add `whenInitialized()` to `InternalProjectInitializer`
  - Wait for the future to complete before starting a mirror migration
    job

Result:

The bug will not affect users as the new mirror format hasn't released.