Fix a bug where new changes in a file editor isn't updated
Motivation:

When a file content is edited and a commit button is clicked, the result
isn't correctly reflected to the push content.

There is a time issue between a change event of `editorRef.current.getValue()`
and a submit event.

Modifications:

- Modified to directly read `editorRef.current.getValue()` when a commit
  message is made.
- Added a filter button to see projects created by a login user.

Result:

Changes in an editor are correctly updated to a commit message in the UI.