Include the project role of a request token or user in project API
Motivation:

I think it is necessary to filter projecs to which users belong or
display their roles in the screen.

Currently, a user information couldn't be obtained from
`/api/v1/projects` API. So we need to call `/api/v1/projects/${projectName}`
to get the metadata. That means 20 API calls should be made to list 20
projects in `/app/projects`. It will cause additional loads to servers
and increase page loading time.

To resolve the N+1 query problem, I propose to add a role of the
requested user to `ProjectDto`.

Modifications:

- Add `RepositoryListener` abstraction that can be continuously updated
  when a new change is pushed to a `Repository`. `Repository.watch()`
  has a similar function but it can be updated only once.
- Add `ProjectMetadata` to `DefaultProject` and use `RepositoryListener`
  to update `ProjectMetadata` in a project when new changes in
  `metadata.json`
- Expose user roles in the Projects page of the new web application.
- Refactor `ControlPlanePlugin` with `RepositoryListener` to simply the
  code

Result:

- The project role of a request token or user are now exposed in the
  project API.
- Users can see their roles in the projects page.
- Users can filter projects by their roles in the project page.