Set up e2e with Playwright (#1001)
Motivation:

E2E tests are important to verify the correctness of behavior. Mocking
tests are fast and easy to inject inputs but they can't check changes in
other sides.

It will not be easy to test all functions with e2e tests. So I am going
to write e2e tests for core features of the webapp.

Modifications:

- Set up Playwright for e2e tests of the webapp