Reload webpages when refreshed
Motivation:

After #1013 is merged, I found out that a web page failed to reload when
a refresh button is clicked or directly access to `/app/projects`

It is caused by a bug of Armeria `FileService` which returns an empty
file instead of a non-existent file when 1) the target path is a
directory and the path does not end with `/`.

The `FileService` has `.html` as a fallback extension.
If `/app/projects` is requested, I expect `/app/projects.html` to be
returned because `/app/projects` is directory. However, it only works
with `FileSystemHttpVfs` but not `ClassPathHttpVfs`.
```
/app/projects/[projectName].html
/app/projects.html
```

The problem wasn't caught by unit tests because `FileSystemHttpVfs` was
always used. During the tests, The class loader directly loads by the
test resouces from the file system.

Modifications:

- Always return `/webapp/index.html` for `/app/**` paths.
  - Next.js will load the correct page on the client side.

Result:

Fixed a regression where the webapp failed to render deep links.