Optimize Commit Frequency in Kubernetes Endpoint Fetching (#1027)
Motivation:
Previously, the KubernetesEndpointFetchingPlugin created a commit each time an endpoint in the KubernetesEndpointGroup was updated. This approach was inefficient, as it resulted in a large number of commits—one for each endpoint update. By batching these updates and making a single commit after a brief delay, we can significantly reduce the number of commits and improve efficiency.

Modifications:
- Modified the plugin to delay the commit by one second after KubernetesEndpointGroup.addListener() is called.

Result:
- Reduced the number of commits required when updating Kubernetes endpoints.