Expose internal projects to users who have `MEMBER` permission (#1053)
Motivation:

Currently non-admin users can't access `@xds` project so they need to
access the resources only through API. The absence of UI will be quite
inconvenient for users. We plan to create UI for xDS, but we don't know
when it will be ready.

As a workaround, it might be better to view the xDS confiugrations by
browsing UI first and switch to the dedicated UI later. For that I
propose to open internal projects to `MEMBER` users.

Modifications:

- Fixed `ProjectApiManager` to contain internal projects if the login
user is a member of the internal project.
- Fixed REST API services to inject the current user to
`ProjectApiManager` when listing projects.
- As an exception, Thrift API does not allow access to internal
projects.

Result:

Administrators now allow users to access the `@xds` project by
registering them as members.