Provide a way to listen to the events of mirroring (#1057)
Motivation:

Mirroring failure is only recorded as a warning, but there is a need to
handle it in a different way. For example, it can be recorded in metrics
or end-users can be notified immediately.

`MirrorListener` is provided as an extension point to utilize various
events occurring in the mirror.

Modifications:

- Introduce `MirrorListener` whose implementations can be loaded
dynamically via Java SPI.
- `onStart()`, `onComplete()` and `onError()` events are added.
- The default behavior is preserved in `DefaultMirrorListener` which is
only used when no custom `MirrorListener` is configured.

Result:

You can now use `MirrorListener` to listen to `Mirror` events.