Synchronously invoke `RepositoryListener` (#1101)
Motivation:

`RepositoryListener` used `Repository.watch()` to continuously invoke
the callback. Since the callback is not invoked immediately upon commit
completion, the cache cannot guarantee a happen-before relationship with
the write operation.

https://github.com/line/centraldogma/pull/1099#pullrequestreview-2632353089

Modifications:

- Fix `CommitWatchers` to reuse `Watch` added for `RepositoryListener`
  - The watch is removed when `WatchListener.onUpdate()` returns true.
- `recurseWatch()` no longer uses a scheduler to submit a new watch task
when notified.
- `blockingfind()` operation is invoked in the same thread as the commit
to ensure that the callback is invoked before the commit's returned
future.
- Remove `Repository.removeListener()` to simplify the implementations.

Result:

`RepositoryListener` is now immediately invoked when the watching files
are modified.