Prohibit application tokens from accessing repositories with GUEST permission (#1093)
Motivation:

We are aiming to implement stricter and more explicit access control in
Central Dogma.

It has been observed that many tokens are accessing repositories as a
GUEST role without being registered with the project. If the token is
exposed to the outside, an attacker could gain access to all
repositories that allow GUEST access.

To resolve the issue, I propose restricting new tokens from accessing
repositories that have GUEST permission. For backward compatibility, the
existing tokens are still allowed to access repositories with GUEST
permission.

Modifications:

- Add `allowGuestAccess` field to `Token` class to indicate if a token
is allowed to access a repository with the GUEST role.
- If the value is null in `tokens.json`, it defaults to `true` for
backward compatibility.
- Fixed `MetadataService` to prohibit tokens whose `allowGuestAccess` is
false from accessing repositories with the GUEST role.

Result:

New application tokens no longer have access to repositories with the
`GUEST` role.