Abort watching on the initial failure
Motivation:

If the initial fetch of `Watcher` fails with non-`CentralDogmaException`,
it is likely that some configuration is wrong or the file is malformed.
So earily completing the initialValueFuture with the exception rather than
retrying indefinitely would help users to debug the reason.

Modifications:

- Complete `initialValueFuture` exceptionally when:
  - `initialValueFuture` is not completed yet
  - The cause is not `CentralDogmaException`

Result:

`Watcher` client stops without retrying if the first attempt fails with
a non-`CentralDogmaException`.