Do not enter readonly mode on lock timeout (#1149)
Motivation:

A lock could be held by a request for a long time when the size is
large. Although this is not a desirable situation, we don't need to regard this
as a system error. So, failing the request makes more sense rather than 
entering read-only mode.

Modifications:

- Do not stop ZooKeeper replication when a lock can be acquired within
one minute.
  - Change the log level to `WARN` from `ERROR`.
- If a request timeout is configured to be longer than the lock timeout,
`LockAcquireTimeoutException` will be returned.
- Add `ApiRequestTimeoutException` for converting Armeria
`RequestTimeoutException` to a `CentralDogmaException`.
  - Also update the client error factories.

Result:

A lock timeout no longer triggers readonly mode.