Reduce the default maxFrameLength from 10MiB to 1MiB (#1147)
Motivation:

When a Central Dogma server received a large push data of over 4.2 MiB, it was observed that a ZooKeeper lock on the repository was held for more than 1 minute. If a lock is held for more than 1 minute, subsequent write operations to the repository may fail to acquire the lock and the server enters readonly mode.

The original `maxFrameLength` seems too large to be handled safely.

Modifications:

- Reduce the default `maxFrameLength` to 1MiB. users approximately write up to 10,000 lines per file which would be generally enough for a configuration file.

Result:

The default `maxFrameLength` has been reduced to 1 MiB.